Make `Equals` and `NotEquals` type safe
We use generics from Go 1.18 to make the functions
`Equals` and `NotEquals` type safe. This change
makes `Equals<Type>` and `NotEquals<Type>` functions
redundent and reduces them to just being wrappers.

Since `error` is not `comparable` and since `error`
can not be used in an union, an errors can no
longer be passed to the new type-safe functions.
Any code that uses `Equals` or `NotEquals` for
comparing errors now must switch to using
`EqualsError` and `NotEqualsError` respectively.
That makes these changes NOT BACKWARD COMPATIBLE.

More on why interfaces can't be unioned with types
in Go 1.18: https://stackoverflow.com/q/72267243